dev_requirements.txt
-i https://pypi.org/simple
-e ./cdk
attrs==21.4.0
aws-cdk-lib==2.21.1
aws-cdk.aws-lambda-python-alpha==2.19.0a0
boto3==1.21.46
botocore==1.24.46
cattrs==1.10.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12 ; python_version >= '3'
click==8.1.2
colorama==0.4.4 ; python_version > '3.4'
constructs==10.0.124
coverage[toml]==6.3.2
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
future==0.18.2
ghp-import==2.0.2
gitdb==4.0.9
gitpython==3.1.27
identify==2.4.12
idna==3.3
importlib-metadata==4.11.3 ; python_version < '3.10'
iniconfig==1.1.1
isort==5.10.1
jinja2==3.1.1
jmespath==1.0.0 ; python_version >= '3.7' and python_version < '4.0'
jsii==1.57.0
mando==0.6.4
markdown==3.3.6
markupsafe==2.1.1
mccabe==0.6.1
mergedeep==1.3.4
mkdocs-git-revision-date-plugin==0.3.2
mkdocs-material-extensions==1.0.3
mkdocs-material==8.2.10
mkdocs==1.3.0
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.18.1
publication==0.0.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pygments==2.12.0
pymdown-extensions==9.3
pyparsing==3.0.8
pytest-cov==3.0.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-mock==3.7.0
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pyyaml-env-tag==0.1
pyyaml==6.0
radon==5.1.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
smmap==5.0.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
watchdog==2.1.7
xenon==0.9.0
yapf==0.32.0
zipp==3.8.0
